requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
gunicorn==25.1.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.10
whitenoise==6.11.0
sqlparse==0.5.3
tzdata==2025.2
django-allauth==65.7.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
requests==2.32.5
charset-normalizer==3.4.5
idna==3.11
urllib3==2.6.3
certifi==2026.2.25
PyJWT>=2.8.0
cryptography>=42.0.0
djangorestframework==3.15.2
